Validate gorouter, tcp-router and routing-api config on bootstrap VM in pre-start
Make sure the canary fails if they are misconfigured
