+++ packages/gorouter/spec
@@ -13,12 +13,13 @@ files:
   - code.cloudfoundry.org/vendor/code.cloudfoundry.org/clock/*.go # gosub
   - code.cloudfoundry.org/vendor/code.cloudfoundry.org/debugserver/*.go # gosub
   - code.cloudfoundry.org/vendor/code.cloudfoundry.org/go-metric-registry/*.go # gosub
-  - code.cloudfoundry.org/gorouter/*.go # gosub
   - code.cloudfoundry.org/gorouter/accesslog/*.go # gosub
   - code.cloudfoundry.org/gorouter/accesslog/fakes/*.go # gosub
   - code.cloudfoundry.org/gorouter/accesslog/schema/*.go # gosub
   - code.cloudfoundry.org/gorouter/accesslog/schema/fakes/*.go # gosub
   - code.cloudfoundry.org/gorouter/accesslog/syslog/*.go # gosub
+  - code.cloudfoundry.org/gorouter/cmd/config-validator/*.go # gosub
+  - code.cloudfoundry.org/gorouter/cmd/gorouter/*.go # gosub
   - code.cloudfoundry.org/gorouter/common/*.go # gosub
   - code.cloudfoundry.org/gorouter/common/health/*.go # gosub
   - code.cloudfoundry.org/gorouter/common/http/*.go # gosub
